requirements.txt
asgiref==3.2.7
astroid==2.3.3
configparser==5.0.0
Django==3.0.4
djangorestframework==3.11.0
isort==4.3.21
lazy-object-proxy==1.4.3
libxml2-python==2.9.10
mccabe==0.6.1
mysql-connector==2.2.9
mysqlclient==1.4.6
Pillow==7.1.1
pylint==2.4.4
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
typed-ast==1.4.1
wrapt==1.11.2
